+++ CHANGELOG.md
@@ -1,12 +1,16 @@
 # ChangeLog
 
-## unreleased changes
+## 3.0.4
 
 機能追加
  * `g.Collision.intersectEntities()` を追加
    * 同一シーン内の任意のエンティティ同士の矩形が重なっているかどうか判定することができます。
  * `g.Collision.intersectLineSegments()` を追加
    * 線分同士の衝突 (交差) を判定することができます。
+
+仕様変更
+ * `g.Label` の生成時、 fontSize (`g.LabelParameterObject#fontSize`) プロパティを省略可能に
+   * デフォルト値は `g.Label` の生成時に指定された font (`g.LabelParameterObject#font`) プロパティの font.size と同じ値になります
 
 ## 3.0.3
 * `g.Game#_reset()` 実行時に `g.Game#_moduleManager` を初期化する処理を追加